refactor: remove roundabout way to connect to socket (#3965)
Seems like this Resolver stuff is just ... unnecessary? It sounds like
it does DNS resolution (with its notoriously sync-only APO), but this
isn't needed in this case. addr is a `SocketAddr`, ie, a resolved IP.

Note the 1 second timeout: it's the same value that was used by actix:
https://github.com/actix/actix/blob/v0.11.0-beta.1/src/actors/resolver.rs#L382

This is a second attempt at https://github.com/near/nearcore/pull/3961, which was reverted in https://github.com/near/nearcore/pull/3964. I've looked at the code more closely, and noticed that atix hard-codded a one-second timeout here.

This is def a problem in theory -- if the peer just drops packets (instead of actively refusing the connection), the whole peer manager get's stuck in the `.wait(ctx);` call for several minutes. Not sure if this is exactly a problem we've observed thoug!

Note that, even with a timeout, it feels like this code is vulnerable to DOS, as it's pretty easy to freeze the peer manager for one second this way.