feat: infrastructure for reusing memory and store between contracts (#4126)
To prepare for optimized pipelined execution of contracts we need to be able to share reusable parts of state.
This is a preparation PR allowing to share certain parts of the VM state between runs. By default, memory state is turned
off, as we need to ensure that no memory data of one contract execution can be seen in another execution.
Once #4128 will be merged, we could test performance impact of different
sharing practises.

Test plan
---------
cargo test
more specifically tests::contract_preload::test_run_preloaded which tests the behavior of the preloader.