fix(jsonrpc): Validators by epoch_id current epoch is not returning Unknown Epoch anymore (#4270)
I've noticed that if I request `validators` method and pass `epoch_id` in params I'm getting `Unknown Epoch` if epoch I request is current one.

I thinks it has to work so this fix makes it possible.

The cause was in the epoch manager. If we request validators with epoch id we try to use `EpochSummary` which deals with "cached" epochs. But we request ongoing epoch which is not cached yet. So on the `view_client` I compare current epoch id with requested one and in case of match I act like it was a request for latest to trigger `EpochInfoAggregator` instead of `EpochSummary`