fix: wasmer1_default feature actually selects wasmer1 (#4281)
We use this feature in ./runtime/near-vm-runner/src/vm_kind.rs.

So, it should be declared in `vm_runner` Cargo.toml rather than in
`vm_logic`.

This is #4279 submitted against master. 

Not sure how to make sure that similar errors don't happen in the future. The culprit here is that you can have `cfg(feature = "some-typo")`, and you won't see any warnings about this anywhere...