Profile by default (#4185)
Having costs breakdown (how much gas was spend for each operation type)
is helpful for both increasing reliability (as we can eyeball if the
costs "make sense") and developer friendliness (as devs can understand
where all the gas is going).

Currently this is hidden behind a feature flag, but I would love to
argue that we should just always enable it.

Performance cost seems to be negligible: we heap-allocate an array once
per apply state, and actual cost increasse are just counter increments,
they should be dwarfed by the host functions themselves.

Correctness-wise, the code is semi-sensitive. It is executed by runtime,
but doesn't affect the results. So it's enough if it doesn't panic.

# Test Plan

All of existing tests now exercise costs councting code. An additional unit-test is added to smoke-test that costs profiling can't panic due to overflow.