fix: remove unwrap and alloc from bytes string serialization (#4296)
Was going to open up an issue, but there was at least this to fix.

My question is how is this safe to use for `Promise` and return data? Is there some guarantee that it's for example JSON formatted and thus always utf8? This seems like something that would be better serialized as base64 or something else, to allow flexibility in the future (even if restricted now) to be some binary encoding