chore: modify stale bot configuration (#4456)
Change days until stale to be longer (sorry for the opposite suggestion earlier). Also @chefsale my experience so far is that it seems that we don't actually need the bot to automatically close the issue. What we want is for the bot to ping when an issue becomes stale so that we can do some manual triage. Is it possible to configure the bot that way?