feat: one button "run estimator in docker \w qemu"  (#4562)
Today, to get precise gas estimations via instruction counts, one needs
to follow a series of manual steps from emu-cost/README.md.

This PR optimizes part of that. It adds `--docker` argument to the
parameter estimator, which takes care of docker and qemu.

With this argument, the estimator on the host machine launches a docker
container, compiles a copy of itself in docker, and runs that copy via
qemu.

There are a couple of manual steps left:

* running neard init
* running genesis-populate
* picking the specific parameter values we use for estimation

Those will be automated in subsequent PR. The desired end goal is that

```console
$ cargo run -- --docker
```

is the *single* command you need to run the whole estimation process.
At that point, all various bash and python scripts will be removed.