feat(indexer): Extend Indexer Framework with logs on start (#4723)
Add a couple of `info!` logs on the indexer side to see the starting process more clearly as it takes quite a while to start the indexer on `testnet` because of the big genesis file. Additional logs should help to understand that the process is ongoing and nothing has stuck.