fix(chain): fix block ordinal to block hash map (#4793)
There is a subtle bug in how we populate `ColBlockOrdinal`. If there is a fork, then depending on when the fork block arrives, it is possible for the fork block to override the block ordinal (since it has the same ordinal as the block on the canonical chain). This PR fixes this issue by updating block ordinal mapping when the canonical chain changes.

Test plan
-----------
* unit tests
* migrated an archival node and confirmed with @mfornet that the previous issue with bridge would indeed be resolved with this PR.