feat: upgrade to wasmer2 via a protocol feature (#4874)
This feature switch our WASM engine implementation from wasmer 0.* to
wasmer 2.*, brining better performance and reliability.

The implementations should be sufficiently similar for this to not be a
protocol upgrade, but we concervatively do a protocol upgrade to be on
the safe side.

Although wasmer2 is faster, we don't change fees with this protocol
version -- we can safely do that in a separate step.


Test Plan
------------

Add an upgrade test, verify that it passes with and without feature. Verify additionally that it fails with the features if the impl doesn't actually enable wasmer2.