fix: simplify, speed up, and fix a latent bug in load_from_file  (#5019)
While investigating which estimator fees are sensitive to the number of accounts in the database, I've noticed that fee estimation *time* surprisingly depends on this a lot.

load_from_file is a relatively hot operation for the estimator. Current code makes
a lot of small reads (=syscalls) which leaves a lot performance on the
table. Adding BufReader here speeds this up for me from 5 seconds to
800ms on a local benchmark.

The `.take(n).read_to_end` pattern has a problem that it'll silently
ignore file trunctation. Ie, it can read fewer than n bytes. Replace
that with `read_exact` which doesn't have such issue.

While we are at it, let's also re-use key&value memory buffers (I get a
measurable speedup from 80ms to 60ms on ad-hoc benchmark)