make view client use a different runtime object to avoid contention (#5125)
The idea is to use a separate runtime for view client to prevent it from blocking transaction processing. Right now, they share the same `RuntimeAdapter`, which means they share the same `EpochManager` and `ShardTries`, both of which are mostly just an in-memory cache around storage access. The cache is read-blocking, so even though view_client only need to read from the caches, it blocks `client` from accessing or writing to the cache. 
 
This fix makes view_client and client used different `RuntimeAdapter` objects. It works because all the uses in `view_client` on `runtime_adapter` are wrappers around `EpochManager` except for `RuntimeAdapter::query`. And the access to `EpochManager` on access the caches. In other words, `view_client` access do not depend on the internal state of `RuntimeAdapter` or `EpochManager`.

Test Plan:
- CI
- Nayduck