Refactor PeerManager to group messages under one enum (#5141)
Closes #5142

We are going to group all messages in `PeerManager` under one enum called `PeerManagerMessageRequest`. `PeerManager` will have one single `Handler<...>` for messages. We will be able to extend this enum with extra paypoad that track sizes of all messages not processed yet, but coming from w given peer, which is needed for (https://github.com/near/nearcore/pull/4927), and make mocking PeerManager easier for test in the future.