runtime: prefer reporting GasLimitExceeded (#5187)
If during execution of a contract both prepaid gas and max burnt gas
limits were exceeded, the `GasCounter` would report `GasExceeded`
error.  This was purely based on which condition was checked first
and entirely accidental based on how the code was written.

Unfortunately, that’s exactly the opposite preference to what we did
in old code which preferred reporting `GasLimitExceeded` instead.

Change the code to match the old behaviour and thus avoid protocol
changes.