Increase cache for RockDB by column (#5212)
We recently found out that our rockdb is not efficient enough.
Therefore we would like to increase cache size of specific columns.
In this PR we will increase cache size for `DBCol::ColState` to 128mb.

We can adjust that cache size later, but we try to be conservative in memory usage. 
Some testing should be done to determine optimize cache sizes. 


Closes https://github.com/near/nearcore/issues/5203