chore: don't include near-vm-runner-standalone in release artifacts (#5299)
We used to use near-vm-runner-standalone as a way to run contracts
locally, but that never worked properly, as the standalone runner can't
perfectly recreate proper blockchain environment. Today, the task of
running a contract "locally" is solved by a sandbox node

https://github.com/near/sandbox

We still keep the code in the repo, as it is sometimes useful for ad-hoc
investigations by contract runtime team. We just forgot to remove it
from releases, which we should have done ages ago.