feat: stabilize LowerRegularOpCost2 (#5365)
# Feature to stabilize

Lower wasm opcode cost from 2_207_874 to 822_756. See https://github.com/near/nearcore/commit/ae9241cef3993832c1e046e328ad576757b42a85 for details, but the main contributor here is that we use intrinsics-based gas counter implementation, which preserves semantics, but makes things faster. 

# Testing and QA

The overall approach is to make gas counting faster, while preserving the semantics exactly, so we need only to verify that behavior stays the same. New gas counter implementation passes all our tests, and was fuzzed against the old implementation. Additionally, during the implementation we introduced a suite of new unit-tests to cover previously uncovered gas counter behavior. We haven't yet run the history replay tool, that's on todo list at the moment. The magnitude of improvement seems reasonable, and matches our back-of-the-envelope calculations (comparing with gas conting completely disabled). Additionally, we still have 3x safety multiplier for the wasm cost. 

It's important to note that what this PR stabilizes is only the cost reduction. The change to the implementation of gas counter is enabled unconditionally, and is where the most risk lies. 

## Premortem

Suppose that we stabilized this feature, and things did go wrong. What could be the issue?

* Gas intrinsification fails at compile time with some kind of panic. 
* Gas intrinsification fails at runtime with some kind of UB.
* Behavior of gas counter is different for some edge cases.
* Something breaks at the point where we integrate wasmer-side gas counter with nearcore


# Checklist
- [ ] Link to nightly nayduck run (`./scripts/nayduck.py -t nightly/nightly.txt`, [docs](https://github.com/near/nearcore/blob/master/nightly/README.md#scheduling-a-run)):  http://nayduck.near.org/
- [x] Update CHANGELOG.md to include this protocol feature in the `Unreleased` section.