feat: Make `deepsize` import and `DeepSizeOf` derive optional  (#5373)
We would like to use `deepsize` crate and `DeepSizeof` derive in our code. In discussed that it's better not to add extra dependencies to our code base if possible. That's useful for external projects using `core/primitives`.

Here is a proposal on how to include `deepsize`/`DeepSizeOf` inside our repo as optional dependency.

We could do it this the same way for other dependencies like, for example:
```
use serde::{Deserialize, Serialize};
use borsh::{BorshDeserialize, BorshSerialize};
```



@frol I hope this solves your concerns. @matklad Can you take a look?