feat: Add `DeepSizeOf` decorator to `PeerManagerMessageRequest` (#5191)
Hey, we are adding a big feature to around database. We will add a functionality to help investigate sizes of all data structures we use. This will require use to add `DeepSizeOf` to code that uses `BorshSerialize`. 

However, this has a lot of benefits:
- rate limiting - this will allow us to control memory usage - make sure our node doesn't ran out of memory
- debugging - it will be easy to check how big given data structure is, and optimize it by size
- network stats - we will be able to generate network stats. How big messages are, what's taking up memory / bandwidth. This will help us analyze, for example why nodes are out of sync sometimes.

This PR only adds `DeepSizeOf` decorator; this shouldn't effect current code execution in anyway.

This PR is  required to implement rate limiting, we want to control maximum memory usage in our node, per each connected peer. The goal is to stop reading from network, when we reach memory limits. https://github.com/near/nearcore/issues/5155
In the future, we will add more use cases this feature as well.

Split: 
- [x] part 1 - make `DeepSizeOf` decoration optional #5373


It will become optional in:
- near-network
- jsonrpc-adversarial-primitives
- near-network-primitives