test: fixing 3 nayduck tests (#5418)
restart.py -- had a flaky assumption that after startup, the current head is <= 2 (probably used to work when our node creation a lot faster)
rpc_finality - inside the code, it started 3 nodes, killed one of them and tried making the transaction, that didn't succeed. AFAIK you need now to have more than 2/3 of the stake for the chain to approve it. So I've changed number of nodes to 4.
sync_chunks_from_archival - was checking that the data was read from the hardcoded archival node id (for example, that first 10 blocks would be read from 2, next 10 blocks from 3 etc) - AFAIK we don't enforce this in the code - so I've updated the test to verify that each block is requested from either node.