fe: Move `ActixMessageWrapper` from `near-network` to `near-rate-limiter` (#5676)
We want to use `near-rate-limiter` across multiple crates. That implies, that we would have to use `ActixMessageWrapper` across mujltiple crates. However, `ActixMessageWrapper` currently is located in `near-network`. And we don't want other crates, to be directly importing from `near-network`.

I propose, we move `ActixMessageWrapper` to the same crate as `ThrottledFramedReader`, that is inside ` near-rate-limiter`.
Both `ActixMessageWrapper`, `ThrottledFramedReader` are meant to be used together, and it makes sense for them to be together.

Later, we may consider moving `near-rate-limiter`, to a new place. However, it's hard to say exactly where. 
Anyway, moving those two structs together, is definitely a good change. Let's do it now.

This is required for the next step. The plan will be added to https://github.com/near/nearcore/issues/5672.