change request_partial_encoded_chunk to filter peers that are up to date (#5636)
Currently, when a node requests chunk parts, it either request from a block producer or one of its peers. Right now, it doesn't check the peer height, so it's possible that this peer is behind and does not have the part, which means the node has to resend the request after 400ms. This change fix that.