logs: add a bit more info in Block sync debug logs (#5934)
This makes two changes to try and provide a bit more debugging
info. One is that we show which peer was chosen to requset a block
from in the block sync debug message. The other is that we also warn
if there were no suitable peers, since otherwise the logs could
possibly be misleading if for example we're connected to 3 peers, none
of which are archival, we're trying to sync from a chain that's > 5
epochs behind and we see a debug message saying:

sync: Block sync: 12345/123456 requesting block ABCDE from 3 peers

when really we're not about to send any request at all.

Issue: https://github.com/near/nearcore/issues/4422