neard: warn when running unsupported build on mainnet or testnet (#5890)
Building via `cargo build` or even `cargo build --release` produces
insufficiently-optimised build.  The recommended way to create an
executable to run in production is `make release` which enables LTO.

There is, however, no mechanism which prevents users from or warns
them about running badly-optimised nodes.

Add a mechanism which will print warnings to the logs when trying to
initialise or run a mainnet or testnet log using an executable which
was not built via `make release` or `make nightly-release`
respectively.